+++ examples/scenarios/serverless-gateway.tf
@@ -1,11 +1,26 @@
 # Don't forget to create auth.tf to configure the provider
 # (see examples/scenarios/_auth.tf for an example)
 
+# Configure a serverless gateway control plane
 resource "konnect_gateway_control_plane" "tfdemo" {
   name         = "Serverless TF Control Plane"
   description  = "This is a sample description"
-  cluster_type = "CLUSTER_TYPE_SERVERLESS"
+  cluster_type = "CLUSTER_TYPE_SERVERLESS_V1"
   auth_type    = "pinned_client_certs"
+  cloud_gateway = true
+}
+
+# Configure a serverless gateway configuration to create the dataplane
+resource "konnect_cloud_gateway_configuration" "my_cloudgatewayconfiguration1" {
+  control_plane_geo = "us"
+  control_plane_id  = konnect_gateway_control_plane.tfdemo.id
+  dataplane_groups = [
+    {
+      provider = "aws"
+      region   = "us"
+    }
+  ]
+  kind    = "serverless.v1"
 }
 
 # Configure a service and a route that we can use to test
@@ -36,16 +51,7 @@ resource "konnect_gateway_data_plane_client_certificate" "my_cert" {
   control_plane_id = konnect_gateway_control_plane.tfdemo.id
 }
 
-resource "konnect_serverless_cloud_gateway" "my_scgw" {
-  control_plane = {
-    id     = konnect_gateway_control_plane.tfdemo.id
-    prefix = replace(replace(konnect_gateway_control_plane.tfdemo.config.control_plane_endpoint, "https://", ""), ".us.cp0.konghq.com", "")
-    region = "us"
-  }
-  cluster_cert     = file("./tls.crt")
-  cluster_cert_key = file("./tls.key")
-}
-
+# Output the endpoint of the serverless gateway
 output "scgw_endpoint" {
-  value = konnect_serverless_cloud_gateway.my_scgw.gateway_endpoint
+  value = konnect_gateway_control_plane.tfdemo.config.control_plane_endpoint
 }