+++ internal/sdk/models/shared/createappauthstrategyresponse.go
@@ -172,6 +172,11 @@ type AppAuthStrategyOpenIDConnectResponse struct {
 	CreatedAt time.Time `json:"created_at"`
 	// An ISO-8601 timestamp representation of entity update date.
 	UpdatedAt time.Time `json:"updated_at"`
+	// Indicates whether this auth strategy supports multiple credentials.
+	// - `true` for Key Auth strategies and when supported for Client Credentials strategies
+	// - `false` when not supported for Client Credentials strategies
+	//
+	SupportsMultipleCredentials *bool `json:"supports_multiple_credentials,omitempty"`
 }
 
 func (a AppAuthStrategyOpenIDConnectResponse) MarshalJSON() ([]byte, error) {
@@ -253,6 +258,13 @@ func (a *AppAuthStrategyOpenIDConnectResponse) GetUpdatedAt() time.Time {
 		return time.Time{}
 	}
 	return a.UpdatedAt
+}
+
+func (a *AppAuthStrategyOpenIDConnectResponse) GetSupportsMultipleCredentials() *bool {
+	if a == nil {
+		return nil
+	}
+	return a.SupportsMultipleCredentials
 }
 
 type AppAuthStrategyKeyAuthResponseStrategyType string
@@ -416,6 +428,10 @@ type AppAuthStrategyKeyAuthResponse struct {
 	CreatedAt time.Time `json:"created_at"`
 	// An ISO-8601 timestamp representation of entity update date.
 	UpdatedAt time.Time `json:"updated_at"`
+	// Indicates whether this auth strategy supports multiple credentials.
+	// Always `true` for KEY_AUTH.
+	//
+	SupportsMultipleCredentials *bool `default:"true" json:"supports_multiple_credentials"`
 }
 
 func (a AppAuthStrategyKeyAuthResponse) MarshalJSON() ([]byte, error) {
@@ -497,6 +513,13 @@ func (a *AppAuthStrategyKeyAuthResponse) GetUpdatedAt() time.Time {
 		return time.Time{}
 	}
 	return a.UpdatedAt
+}
+
+func (a *AppAuthStrategyKeyAuthResponse) GetSupportsMultipleCredentials() *bool {
+	if a == nil {
+		return nil
+	}
+	return a.SupportsMultipleCredentials
 }
 
 type CreateAppAuthStrategyResponseType string